Add support for multiple devices
Add a new service mapping to better support multiple devices. The
new mapping allows us to get rid of the serviceId-based overhead
such as subscribeService and publishService, as well as the hardcoded
SystemConfiguration.services. Now all the available dbus services
are maintained in a single place, and can be accessed on-demand.