Disable system relay if it is not on manual mode
Disable the usage of system relay 0 if it is set on any other mode than
manual. An approppriate warning and info on how to change the relay function
is shown below the relay select in this case. Trying to trigger a
previously deployed relay with any other mode than manual will show a
warning message on Node-RED's debug panel. Restructure the relay service
logic to get rid of regular expressions. Use more functional style.