Improve 'no devices found' text
A warning text that was shown when there were no connected
(corresponding) devices found was ambigous. Modify the edit form to show
disabled selects and a friendly warning text when this dbus service is
not connected. The user is guided to either try out a different node or
check the device connection.