Fix for multiple items in ItemsChanged array
In case of an ItemsChanged package, only the last of the items that was in the received array was being parsed correctly. This change fixes that and parsers all items from the array, resolving issue #145.
Thus far this bug had gone unnoticed as there are several nodes that have such few changes per second that the ItemsChanged array often consists of a single item. Each item is also polled every 5 seconds (pollInterval), which was always parsed correctly.