Use solar forecast for today to switch a relay
This example shows what you can do with solar forecasting.
The relay will be set to open when the predicted PV yield of
today is above a threshold. The threshold is 1000 in this
example.

This could be a crude way to only use power when there is
enough solar predicted.