Add new virtual devices
Underlying dbus-victron-virtual library update to version 0.1.4,
which adds new functionality. Now calling addSettings before creating the service on the dbus. This allows for getting a correct DeviceInstance _before_ the service gets created. Allowing for quite a big cleanup of the code.
Also adding new virtual devices: battery and pv inverter.

And added the option to set default values for the node, which will result in the device to appear on VRM, without you having to write all needed values first.