Use configuration node for polling
Instead of using the environment variable, use the configuration client for setting the behavior, defaulting to not polling.
This makes for a better user experience as it saves him from having to alter the envronment via ssh.
Do note that a re-deploy won't alter the behavior after changing the checkbox and storing. A restart of Node-RED is needed for that.
This is acceptable, as most users never should have to revert tothe old behavior.

Also make the checking more robust by making sure that we are connected before retrieving a value.