VictronDbusListener reconnecting, do not fail _requestRoot on virtual devices
On reconnect, there is a race condition between (a) the virtual node defining
its interfaces with dbus, and (b) the listener requesting all roots known in
its cache: If _requestRoot() comes first, the dbus call GetItems() fails, as
the interface does not exist on dbus (yet).

We "fix" this by not failing, if we detect in _requestRoot() by the service
name that it's a virtual device. This is not bullet-proof, as there may be a
virtual device defined outside of the node-red process, and in this case we
*should* fail. The possible impact is that we would have stale data in the
cache of VictronDbusListener.