nodes of type 'victron-virtual' now handle one input, with a payload of one or more properties to change
One critical change which needs more testing / verification if it is actually
correct: on a 'close' event, the dbus connection is now ended
*unconditionally*. Previously, the connection was ended only when the node was
the last in the set 'nodeInstances'. The old behavior seems inconsistent: We
always create a new dbus connection for a victron-virtual node, no matter how
many other node instances there are, but on a 'close' event we do not always
close the connection; This may be a leftover of some inconsistent refactoring.

The new behavior seems to fix a bug where sending changed properties to a
'victron-virtual' node worked well (including an 'ItemsChanged' dbus signal),
but did not work any more after a 'deploy' of the flow.