virtual nodes, fix race condition between receiving input and initialising dbus
Previously, 'node.setValuesLocally' might not be available yet on receiving
input to the node. Initialising setValuesLocally is part of setting up the dbus
connection (which is async).

We now cater for this case, where input arrives before completion of the
initialization: If not available yet, we queue the input, and on availability
we call 'node.setValuesLocally()' for anything queued.