Resolve killing the dbus connection
By calling getValue wrongly, you can accidentally kill the dbus connection. If it is called with an interface name that does not exist, it terminates the dbus connection. This patch introduces searching in the known services list for the right destination when the getValue function is called with a `com.victronenergy.temperature/${deviceInstance}` way.