Improve edit panel layout
Previously the layout of the edit panel wasn't too UX friendly. These
changes help in making it more userfriendly by aligning everything
better.
The switches don't appear on VRM, so hide the "initialize with default
values" when the switch is selected.

Note, this builds on top of the faberd/virtual-device-only-1-switch-instead-of-4
branch