Probably fixes group / custom name config issue
While testing, group name and custom name of a switch node were not always
adjustable. This commit should fix it by disabling persistence for (hopefully
all relevant) configuration values: Configuration values are persisted as part
of the flow, so it should be unnecessary (actually, causing unwanted behavior)
to persist them.